# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Server",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:9501",
            "description": "Local server"
        },
        {
            "url": "https://catchupai.on-forge.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/v1/custom-prompts": {
            "get": {
                "tags": [
                    "CustomPrompts"
                ],
                "summary": "List the current user custom prompts",
                "operationId": "api.v1.custom-prompts.index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CustomPromptResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CustomPrompts"
                ],
                "summary": "Create a custom prompt",
                "operationId": "api.v1.custom-prompts.store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "學習筆記摘要"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "請以學習筆記的風格整理這部影片的重點…"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomPromptResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/custom-prompts/{promptId}": {
            "get": {
                "tags": [
                    "CustomPrompts"
                ],
                "summary": "Show one custom prompt",
                "operationId": "api.v1.custom-prompts.show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/promptId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomPromptResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CustomPrompts"
                ],
                "summary": "Replace a custom prompt",
                "operationId": "api.v1.custom-prompts.update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/promptId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "學習筆記摘要"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "請以學習筆記的風格整理這部影片的重點…"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomPromptResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CustomPrompts"
                ],
                "summary": "Delete a custom prompt",
                "operationId": "api.v1.custom-prompts.destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/promptId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/media/{mediaId}/chat/sessions/{sessionId}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Get a chat session with its messages",
                "description": "GET /v1/media/{mediaId}/chat/sessions/{sessionId}.\n\n回傳指定 ChatSession 與其訊息記錄。",
                "operationId": "api.v1.media.chat.sessions.show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/mediaId"
                    },
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01jsvgt3prpypqwex4wj78bznk"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatSessionDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Media"
                ],
                "summary": "Delete a chat session",
                "description": "DELETE /v1/media/{mediaId}/chat/sessions/{sessionId}.\n\n刪除指定 ChatSession（soft delete）。",
                "operationId": "api.v1.media.chat.sessions.destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/mediaId"
                    },
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01jsvgt3prpypqwex4wj78bznk"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/rss/{rssId}": {
            "delete": {
                "tags": [
                    "RSS"
                ],
                "summary": "Unsubscribe from RSS feed",
                "operationId": "api.v1.rss.destroy",
                "parameters": [
                    {
                        "name": "rssId",
                        "in": "path",
                        "description": "RSS feed ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01JCXYZ123456789ABCDEFGHIJ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "404": {
                        "description": "RSS feed not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/sources/{sourceId}": {
            "get": {
                "tags": [
                    "Sources"
                ],
                "summary": "Get details of a single source",
                "operationId": "api.v1.sources.show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/sourceId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "01JCXYZ123456789ABCDEFGHIJ"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Google Developers"
                                        },
                                        "url": {
                                            "type": "string",
                                            "format": "url",
                                            "example": "https://www.youtube.com/channel/UC295-Dw4tzbkl8M9I2GFRtg"
                                        },
                                        "type": {
                                            "type": "string",
                                            "enum": [
                                                "channel",
                                                "playlist"
                                            ],
                                            "example": "channel"
                                        },
                                        "notify": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "thumbnail": {
                                            "type": "string",
                                            "example": "https://yt3.ggpht.com/...",
                                            "nullable": true
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "News and tutorials from the Google Developers team."
                                        },
                                        "subscriber_count": {
                                            "type": "integer",
                                            "example": 1230000,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sources"
                ],
                "summary": "Update notification setting for a subscribed source",
                "operationId": "api.v1.sources.update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/sourceId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "notify"
                                ],
                                "properties": {
                                    "notify": {
                                        "description": "Enable or disable email notification",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Source updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SourceResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sources"
                ],
                "summary": "Unsubscribe from a source",
                "operationId": "api.v1.sources.destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/sourceId"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/subscriptions/{subscriptionId}": {
            "put": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Confirm subscription after successful Paddle payment",
                "operationId": "api.v1.subscriptions.update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/subscriptionId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transaction_id"
                                ],
                                "properties": {
                                    "transaction_id": {
                                        "description": "Paddle transaction ID",
                                        "type": "string",
                                        "example": "txn_01abc..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Cancel subscription",
                "description": "取消訂閱.",
                "operationId": "api.v1.subscriptions.destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/subscriptionId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomPromptResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01k9v7m2q8n4r6t0w3y5z7b1c9"
                    },
                    "title": {
                        "type": "string",
                        "example": "學習筆記摘要"
                    },
                    "content": {
                        "type": "string",
                        "example": "請以學習筆記的風格整理這部影片的重點…"
                    },
                    "model": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/AiModelResource"
                            }
                        ],
                        "nullable": true,
                        "description": "Only present when eager loaded. Null when the config has no model pinned."
                    },
                    "sources": {
                        "description": "Channels / playlists this config applies to. Only present when eager loaded.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SourceResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "401": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "422": {
                "description": "Invalid request parameters",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "messages": {
                                    "description": "驗證失敗的實際回應。\n\n`App\\Exceptions\\InvalidRequestException::render()` 回的是 422 加上 `messages`\n這個鍵（不是 400、也不是 `errors`）——`$statusCode` 與 `render()` 兩處都寫死了。\n既有端點多半沿用 Http400，那份與執行期行為對不上；新端點請用這一份。",
                                    "type": "object",
                                    "example": {
                                        "redirect": [
                                            "導轉網址為必填。"
                                        ]
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
...
    },
...
}